+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix alert url since we changed the Mimir Datasource name.
+
 ## [0.66.0] - 2026-03-17
 
 ### Changed