+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.66.1] - 2026-03-19
+
 ### Fixed
 
 - Fix alert url since we changed the Mimir Datasource name.
@@ -996,7 +998,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.66.0...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.66.1...HEAD
+[0.66.1]: https://github.com/giantswarm/observability-operator/compare/v0.66.0...v0.66.1
 [0.66.0]: https://github.com/giantswarm/observability-operator/compare/v0.65.0...v0.66.0
 [0.65.0]: https://github.com/giantswarm/observability-operator/compare/v0.64.0...v0.65.0
 [0.64.0]: https://github.com/giantswarm/observability-operator/compare/v0.63.0...v0.64.0